set env. c1, to fast get c1.png

+++ optc.sh
@@ -5,6 +5,14 @@ PKG=$1
 ID=$2
 SUBDIR=$3
 
+if [ -n "$c1" ]; then
+  adb shell ls /data/data/$PKG/files/Cache/GNPCACHE/*/*c1.png | tr '\r' ' ' | xargs -n1 adb pull
+  if [ $? == 0 ]; then
+    mv *.png $BASEDIR'/png/'$SUBDIR'/'
+  fi
+  exit 0
+fi
+
 # pull GNPCACHE, mv *.png to SUBDIR
 rm -rf GNPCACHE
 adb pull /data/data/$PKG/files/Cache/GNPCACHE
